+++ CHANGELOG.md
@@ -1,6 +1,9 @@
 # Changelog
 Notable changes to this project.
 
+## [dev]
+- add deprecation warning
+
 ## [2.17.0] - 2024-02-18
 - signals: add 'list_datasets' and 'download_dataset'